llm-apps/job-scraper/pyproject.toml
[project]
name = "job-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.5",
    "boto3>=1.42.83",
    "fastapi>=0.135.3",
    "gradio>=5.0.0",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "python-decouple>=3.8",
    "python-multipart>=0.0.22",
    "requests>=2.33.1",
    "starlette>=1.0.0",
    "uvicorn>=0.43.0",
]
